+++ hive/lib/src/box/lazy_box.dart
@@ -11,7 +11,7 @@ abstract class LazyBox extends Box {
   Future<dynamic> get(dynamic key, {dynamic defaultValue});
 
   @override
-  Future<dynamic> getAt(int index, {dynamic defaultValue});
+  Future<dynamic> getAt(int index);
 
   /// Not supported for lazy boxes.
   @override