+++ hive/test/integration/different_frame_types_test.dart
@@ -25,5 +25,5 @@ void main() {
     test('normal box', () => _performTest(false));
 
     test('lazy box', () => _performTest(true));
-  });
+  }, timeout: longTimeout);
 }