+++ hive/test/integration/integration.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:hive/hive.dart';
 import 'package:hive/src/box/box_base.dart';
 import 'package:hive/src/hive_impl.dart';
@@ -12,13 +14,14 @@ Future<Box> openBox(bool lazy) async {
     var dir = await getTempDir();
     hive.init(dir.path);
   }
-  return await hive.openBox('box', lazy: lazy, crashRecovery: false);
+  var id = Random().nextInt(99999999);
+  return await hive.openBox('box$id', lazy: lazy, crashRecovery: false);
 }
 
 Future<Box> reopenBox(Box box) async {
   await box.close();
   var hive = (box as BoxBase).hive;
-  return await hive.openBox('box', lazy: box.lazy, crashRecovery: false);
+  return await hive.openBox(box.name, lazy: box.lazy, crashRecovery: false);
 }
 
 const longTimeout = Timeout(Duration(minutes: 2));