+++ hive/test/integration/put_many_simultaneously_test.dart
@@ -31,5 +31,5 @@ void main() {
     test('normal box', () => _performTest(false));
 
     test('lazy box', () => _performTest(true));
-  });
+  }, timeout: longTimeout);
 }